int64 fixes

+++ core/variant_op.cpp
@@ -3401,8 +3401,8 @@ void Variant::interpolate(const Variant &a, const Variant &b, float c, Variant &
 		}
 			return;
 		case INT: {
-			int va = a._data._int;
-			int vb = b._data._int;
+			int64_t va = a._data._int;
+			int64_t vb = b._data._int;
 			r_dst = int((1.0 - c) * va + vb * c);
 		}
 			return;
